move constants.mjs into common folder
