+++ components/discord_bot/actions/create-channel-invite/create-channel-invite.mjs
@@ -1,4 +1,4 @@
-import constants from "../../constants.mjs";
+import constants from "../../common/constants.mjs";
 import common from "../common.mjs";
 
 export default {
