+++ components/discord_bot/actions/create-guild-channel/create-guild-channel.mjs
@@ -1,5 +1,5 @@
 import common from "../../common.mjs";
-import constants from "../../constants.mjs";
+import constants from "../../common/constants.mjs";
 
 const { discord } = common.props;
