+++ components/discord_bot/actions/list-channel-invites/list-channel-invites.mjs
@@ -1,4 +1,4 @@
-import constants from "../../constants.mjs";
+import constants from "../../common/constants.mjs";
 import common from "../common.mjs";
 
 const { discord } = common.props;
