+++ components/discord_bot/common/constants.mjs
