+++ components/discord_bot/sources/common.mjs
@@ -1,5 +1,5 @@
-import constants from "../constants.mjs";
 import common from "../common.mjs";
+import constants from "../common/constants.mjs";
 
 export default {
   ...common,
